refactor: common fn for addAll
